refactor: dto 폴더 및 timeline 코드 정리
